Added 'watching' more flag to error handler module
