--- gulp/common/handle-error.js
@@ -1,11 +1,0 @@
-handleError = function(err) {
-  var displayErr = gutil.colors.red(err);
-  if (watching) {
-    gutil.log(displayErr);
-    gutil.beep();
-    this.emit('end');
-  } else {
-    throw displayErr;
-  }
-}
-module.exports = handleError
