+++ gulp/tasks/scripts-coffee.js
@@ -1,6 +1,6 @@
 var gulp = require('gulp');
 var $ = require('gulp-load-plugins')();
-var handleError = require('../common/handle-error');
+var errorHandler = require('../common/error-handler');
 var coffee = require('gulp-coffee');
 
 gulp.task('scripts-coffee', function() {
@@ -8,7 +8,7 @@ gulp.task('scripts-coffee', function() {
     .pipe($.sourcemaps.init())
     .pipe(coffee({
       bare: true
-    }).on('error', handleError))
+    }).on('error', errorHandler.process))
     .pipe($.sourcemaps.write('./maps'))
     .pipe(gulp.dest('.tmp/scripts'))
     .pipe($.size());
