+++ gulp/tasks/styles-less.js
@@ -1,14 +1,14 @@
 var gulp = require('gulp');
 var $ = require('gulp-load-plugins')();
 var less = require('gulp-less');
-var handleError = require('../common/handle-error');
+var errorHandler = require('../common/error-handler');
 
 gulp.task('styles-less', function() {
   return gulp.src('app/styles/main.less')
     .pipe($.sourcemaps.init())
     .pipe(less({
       paths: ['.']
-    }).on('error', handleError))
+    }).on('error', errorHandler.process))
     .pipe($.autoprefixer('last 1 version'))
     .pipe($.sourcemaps.write('./maps'))
     .pipe(gulp.dest('.tmp/styles'))
